requirements.txt
Flask==2.3.3
google-generativeai==0.5.0
yfinance==0.2.36
pandas==2.2.2